+++ packages/uql-orm/src/maria/mariadbQuerier.test.ts
@@ -1,5 +1,6 @@
+import { expect } from 'vitest';
 import { AbstractSqlQuerierIt } from '../querier/abstractSqlQuerier-test.js';
-import { createSpec } from '../test/index.js';
+import { createSpec, TaxCategory } from '../test/index.js';
 import { MariadbQuerierPool } from './mariadbQuerierPool.js';
 
 export class MariadbQuerierIt extends AbstractSqlQuerierIt {
@@ -17,6 +18,24 @@ export class MariadbQuerierIt extends AbstractSqlQuerierIt {
       }),
       'INT AUTO_INCREMENT PRIMARY KEY',
     );
+  }
+
+  override async shouldUpsertOne() {
+    const pk = '507f1f77bcf86cd799439011';
+
+    const insertResult = await this.querier.upsertOne(TaxCategory, { pk: true }, { pk, name: 'Some Name C' });
+    expect(insertResult.firstId).toBeDefined();
+    expect(insertResult.created).toBeUndefined();
+
+    const record2 = await this.querier.findOne(TaxCategory, { $select: { name: true }, $where: { pk } });
+    expect(record2).toMatchObject({ name: 'Some Name C' });
+
+    const updateResult = await this.querier.upsertOne(TaxCategory, { pk: true }, { pk, name: 'Some Name D' });
+    expect(updateResult.firstId).toBeDefined();
+    expect(updateResult.created).toBeUndefined();
+
+    const record3 = await this.querier.findOne(TaxCategory, { $select: { name: true }, $where: { pk } });
+    expect(record3).toMatchObject({ name: 'Some Name D' });
   }
 }